added output.txt

+++ image-pt.c
@@ -112,7 +112,7 @@ enum KernelTypes GetKernelType(char* type){
 //main:
 //argv is expected to take 2 arguments.  First is the source file name (can be jpg, png, bmp, tga).  Second is the lower case name of the algorithm.
 int main(int argc,char** argv){
-    long t1,t2;
+    double t1,t2;
     pthread_t *threads;
     
     stbi_set_flip_vertically_on_load(0); 
@@ -164,7 +164,7 @@ int main(int argc,char** argv){
    
    t2 = time(NULL);
 
-    printf("Took %ld seconds\n", t2 - t1);
+    printf("Took %lf seconds\n", t2 - t1);
 
     stbi_write_png("output.png",destImage->width,destImage->height,
 		    destImage->bpp,destImage->data,destImage->bpp*destImage->width);
